+++ haskell/wasm-guest/test/Main.hs
@@ -4,7 +4,8 @@ import ExoMonad.Guest.Tools.Chainlink.PureTest (pureTests)
 import ExoMonad.Guest.Tools.EventsTest (eventsTests)
 import ExoMonad.Guest.Tools.PollWorkersTest (pollWorkersTests)
 import ExoMonad.Guest.Tools.ResumePrTest (resumePrTests)
+import ExoMonad.Guest.Tools.ResumeBlockedLeafTest (resumeBlockedLeafTests)
 import Test.Tasty (defaultMain, testGroup)
 
 main :: IO ()
-main = defaultMain $ testGroup "WASM guest" [pureTests, eventsTests, resumePrTests, pollWorkersTests]
+main = defaultMain $ testGroup "WASM guest" [pureTests, eventsTests, resumePrTests, resumeBlockedLeafTests, pollWorkersTests]